refactor(postgres): extract shared Postgres helpers out of Neon
Move applyMigrations / runSql / listSqlFiles / readSqlFile /
parsePostgresOrigin from packages/alchemy/src/Neon/ to a new shared
packages/alchemy/src/Postgres/ module so non-Neon Postgres-flavoured
providers (next commit: Vultr Managed Postgres) can reuse them without
a cross-vendor import.

The Neon module's own files become thin re-export shims so existing
imports (alchemy/Neon/Migrations, etc.) keep working — no public API
break.
